# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Traveluxe OS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: vip_tier
          in: query
          schema:
            type: string
        - name: inactive
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "201":
          description: Created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Client details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientWithHistory"
    put:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/check-duplicate:
    get:
      operationId: checkClientDuplicate
      tags: [clients]
      summary: Check for duplicate client by WhatsApp or email
      parameters:
        - name: whatsapp
          in: query
          schema:
            type: string
        - name: email
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Duplicate check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DuplicateCheckResult"
  /clients/merge:
    post:
      operationId: mergeClients
      tags: [clients]
      summary: Merge two client profiles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MergeClientsBody"
      responses:
        "200":
          description: Merged client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
...
  /commissions/settlements:
    post:
      operationId: createSettlement
      tags: [commissions]
      summary: Create a commission settlement for a driver
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSettlementBody"
      responses:
        "201":
          description: Settlement created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settlement"
  /commissions/payouts:
    post:
      operationId: createPayout
      tags: [commissions]
      summary: Create a driver payout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePayoutBody"
      responses:
        "201":
          description: Payout created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payout"
...
components:
  schemas:
...
    Client:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        nationality:
          type: string
        language_preference:
          type: string
        vip_tier:
          type: string
        notes:
          type: string
        inactive:
          type: boolean
        created_at:
          type: string
        created_by:
          type: string
        total_bookings:
          type: integer
        total_spent:
          type: number
        last_booking_date:
          type: string
      required: [id, name, whatsapp, vip_tier]
    ClientWithHistory:
      allOf:
        - $ref: "#/components/schemas/Client"
        - type: object
          properties:
            bookings:
              type: array
              items:
                $ref: "#/components/schemas/Booking"
    CreateClientBody:
      type: object
      properties:
        name:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        nationality:
          type: string
        language_preference:
          type: string
        vip_tier:
          type: string
        notes:
          type: string
        inactive:
          type: boolean
      required: [name, whatsapp, vip_tier]
    DuplicateCheckResult:
      type: object
      properties:
        found:
          type: boolean
        client:
          $ref: "#/components/schemas/Client"
        match_type:
          type: string
      required: [found]
    MergeClientsBody:
      type: object
      properties:
        primary_id:
          type: string
        secondary_id:
          type: string
      required: [primary_id, secondary_id]
...
    Booking:
      type: object
      properties:
        id:
          type: string
        tvl_ref:
          type: string
        client_id:
          type: string
        client_name:
          type: string
        client_vip_tier:
          type: string
        quote_id:
          type: string
        service_type:
          type: string
        direction:
          type: string
        pickup:
          type: string
        dropoff:
          type: string
        destination:
          type: string
        flight_number:
          type: string
        date_time:
          type: string
        passengers:
          type: integer
        luggage:
          type: integer
        vehicle_type:
          type: string
        nameboard:
          type: string
        special_requests:
          type: string
        additional_charges:
          type: number
        price:
          type: number
        tvl_commission:
          type: number
        driver_receives:
          type: number
        commission_type:
          type: string
        payment_status:
          type: string
        payment_method:
          type: string
        commission_status:
          type: string
        payout_status:
          type: string
        source:
          type: string
        status:
          type: string
        operator_id:
          type: string
        operator_name:
          type: string
        driver_id:
          type: string
        driver_name:
          type: string
        driver_vehicle:
          type: string
        return_booking_id:
          type: string
        is_amended:
          type: boolean
        notes:
          type: string
        cancellation_reason:
          type: string
        cancellation_fee:
          type: number
        created_at:
          type: string
        updated_at:
          type: string
        created_by:
          type: string
        duration:
          type: number
      required: [id, tvl_ref, service_type, status, price]
...
    Settlement:
      type: object
      properties:
        id:
          type: string
        driver_id:
          type: string
        driver_name:
          type: string
        week_start:
          type: string
        week_end:
          type: string
        total_amount:
          type: number
        settled_at:
          type: string
        settled_by:
          type: string
        notes:
          type: string
      required: [id, driver_id, total_amount]
    CreateSettlementBody:
      type: object
      properties:
        driver_id:
          type: string
        week_start:
          type: string
        week_end:
          type: string
        booking_ids:
          type: array
          items:
            type: string
        notes:
          type: string
      required: [driver_id, week_start, week_end, booking_ids]
    Payout:
      type: object
      properties:
        id:
          type: string
        driver_id:
          type: string
        driver_name:
          type: string
        week_start:
          type: string
        week_end:
          type: string
        total_amount:
          type: number
        paid_at:
          type: string
        paid_by:
          type: string
        notes:
          type: string
      required: [id, driver_id, total_amount]
    CreatePayoutBody:
      type: object
      properties:
        driver_id:
          type: string
        week_start:
          type: string
        week_end:
          type: string
        booking_ids:
          type: array
          items:
            type: string
        notes:
          type: string
      required: [driver_id, week_start, week_end, booking_ids]
...